backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
